+++ candle/groups/frontend.py
@@ -1,0 +1,37 @@
+
+from typing import Dict
+from flask import render_template, Blueprint, request
+
+from candle.groups.models import StudentGroup
+from candle.groups.search import search_groups
+from candle.timetable.blueprints.readonly import register_timetable_routes
+
+groups = Blueprint('groups', __name__, template_folder='templates', url_prefix='/groups')
+
+
+@groups.route('/')
+def listing():
+    """Show all student groups."""
+    groups_list = search_groups(request.args.get("q"))
+    student_groups_dict = get_student_groups_sorted_by_first_letter(groups_list)
+    title = "Rozvrhy krúžkov"
+    return render_template('groups/listing.html', student_groups_dict=student_groups_dict,
+                           title=title, web_header=title)
+
+
+def get_group(group_url_id: str) -> StudentGroup:
+    return StudentGroup.query.filter((StudentGroup.id_==group_url_id) | (StudentGroup.name==group_url_id)).first_or_404()
+
+
+register_timetable_routes(groups, get_group)
+
+
+def get_student_groups_sorted_by_first_letter(student_groups) -> Dict:
+    """Return student-groups in a dictionary sorted by the first letter."""
+    result_dict = {}
+    for group in student_groups:
+        first_letter = group.name[0]
+        if first_letter not in result_dict:
+            result_dict[first_letter] = []
+        result_dict[first_letter].append(group)
+    return result_dict
