+++ candle/models.py
@@ -79,25 +79,6 @@ class Teacher(SchoolTimetable):
         return self.short_name
 
 
-student_group_lessons = db.Table('student_group_lessons',
-                                 db.Column('student_group_id', db.Integer, db.ForeignKey('student_group.id')),
-                                 db.Column('lesson_id', db.Integer, db.ForeignKey('lesson.id')))
-
-
-class StudentGroup(SchoolTimetable):
-    id_ = db.Column('id', db.Integer, primary_key=True)
-    name = db.Column(db.String(30), nullable=False)
-    lessons = db.relationship('Lesson', secondary=student_group_lessons, lazy='dynamic')
-
-    @property
-    def timetable_name(self) -> str:
-        return f"Rozvh krúžku {self.name}"
-
-    @property
-    def timetable_short_name(self) -> str:
-        return self.name
-
-
 class Lesson(db.Model):
     id_ = db.Column('id', db.Integer, primary_key=True)
     day = db.Column(db.Integer, nullable=False)
