+++ candle/panel/templates/panel/panel.html
@@ -41,7 +41,7 @@
         <div class="panel_cast_obsah">
             {% include('search/group_search.html') %}
             <div class="other_links">
-                <a href="{{ url_for('student_group.list_student_groups') }}">Rozvrhy všetkých krúžkov</a>
+                <a href="{{ url_for('groups.listing') }}">Rozvrhy všetkých krúžkov</a>
             </div>
         </div>
     </div>
