+++ candle/timetable/templates/timetable/list.html
@@ -8,7 +8,7 @@
         <td>{{ lesson.day_abbreviated }}</td>
         <td>{{ lesson.start_formatted }}</td>
         <td>{{ lesson.end_formatted }}</td>
-        <td><a href="{{ url_for('room.show_timetable', slug=lesson.room.name) }}">{{ lesson.room.name }}</a></td>
+        <td><a href="{{ url_for('rooms.show_timetable', slug=lesson.room.name) }}">{{ lesson.room.name }}</a></td>
         <td>{{ lesson.type }}</td>
         <td>{{ lesson.subject.short_code }}</td>
         <td>
